==> Scripts_MultiStaning/Step3_Sunburst_Plots.Rmd <==
---
title: "SunBurst_Plots"
author: "Melissa"
date: "27/04/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r directories creation}
#the main Dir should be the root directory in which both input, output and script folders are stored
mainDir= here::here()

mainDir= "C:/Users/Melissa S/Desktop/BRCA1_Tumorigenesis/Tumor_BRCA_p53"

inputDir=  file.path(mainDir,"input", "multistaining") 
outputDir= file.path(mainDir, "output", "multiStaining")
#####################
RDatadir <- file.path(outputDir,"RData") ; if(!file.exists(RDatadir)){dir.create(RDatadir)}
QCdir <- file.path(outputDir, "QC") ; if(!file.exists(QCdir)){dir.create(QCdir)}
PLOTdir <- file.path(outputDir, "Plots") ; if(!file.exists(PLOTdir)){dir.create(PLOTdir)}

```
```{r}
dtest=qs::qread( "C:/Users/Melissa S/Desktop/PROJECTS/BRCA1_Tumorigenesis/REVIEW/output/mouse/multistaining/RData/Table_Occurences_Combinations_changedEM.qs")
```


```{r}
library(ggplot2)
library(ggpubr)
library(dplyr)
library(reshape2)
library(qdapRegex)
library(mclust)
library(stringr)
library(data.table)
library(sunburstR)
```


```{r}
gtheme= theme (axis.text.x = element_text(size=12, face="bold", colour = "black", angle = 0, hjust = 0.5),axis.text.y = element_text(size=12, face="bold", colour = "black"),legend.text = element_text(face = "bold", size = 12),text=element_text(size=14, face="bold", colour = "black"))
###
themplot=theme_bw() + theme(panel.border = element_blank(), panel.grid.major = element_blank(),
panel.grid.minor = element_blank(), axis.line = element_line(colour = "black"),axis.ticks.x=element_blank())
```


```{r}
mat= qs::qread(paste0(RDatadir, "/Final_Matrix_All_Binarized_Markers.qs"))
mat$name=paste0(mat$sample, "_", mat$Parent)
```

First remove stroma; then assign cell types according to marker expression
keep mat unhanged
```{r}
 mat=qs::qread(paste0(RDatadir, "/Thresholded_matrix_Binarized_WoStroma.qs"))
epith=mat[, -c(1, 11:17)]

#epith$name=paste0(epith$sample, epith$Parent)
epith$Epithtype= ifelse(epith$Krt5== 1 & epith$Krt8== 0, "Basal", ifelse(epith$Krt5== 0 & epith$Krt8== 1, "Luminal", ifelse(epith$Krt5== 1 & epith$Krt8== 1, "Krt5Krt8p", "none")))

epith= epith %>% select(! c(  "Krt5", "Krt8"))

epith$Cdkn2a= ifelse(epith$Cdkn2a== 0, "p16neg", "p16pos")

epith$Ki67= ifelse(epith$Ki67== 0, "Ki67neg", "Ki67pos")
```



```{r}
epith$EMstate = ifelse(epith$Ecad== 1 & epith$Ncad== 0 & epith$Vim==0 , "Epithelial",
                       
                       ifelse(epith$Ncad== 1 & epith$Ecad== 0 | epith$Vim== 1 & epith$Ecad== 0 , "Mesenchymal",
                              ifelse(epith$Ncad== 0 & epith$Ecad== 0 & epith$Vim==0, "None",
                              "EMP" )))


epith= epith %>% select(! c("Ecad", "Ncad", "Vim"))


#epith= epith[, c(4,5,1,2,6,3)]
 #create a table of occurences of all combinations:
dtest= as.data.frame(table(epith))

dtest= as.data.frame(table(epith))

qs::qsave(dtest, paste0(RDatadir, "/Table_Occurences_Combinations_changedEM.qs"))
dtest=qs::qread(paste0(RDatadir, "/Table_Occurences_Combinations_changedEM.qs"))
```












Make a ggplot version:
```{r}
sunburst_ggplot_multiplelayers= function(dt){
  
rectdata <- dt %>%
  group_by( Epithtype ) %>%
  summarise( tot_cells = sum( Freq ) ) %>%
  ungroup() %>%
  mutate(
    ymax = cumsum( tot_cells ),
    ymin = lag( ymax, n = 1, default = 0 )
  )


outerCircleData1 <- dt %>%
  group_by( Epithtype, Cdkn2a ) %>%
  summarise( tot = sum( Freq ) ) %>%
  left_join( rectdata %>% select( Epithtype,tot_cells) ) %>%
  ungroup() %>%
  mutate(
    ymax = cumsum(tot),
    ymin = lag( ymax, n = 1, default = 0 )
  ) 

outerCircleData2 <- dt %>%
  group_by( Epithtype, Cdkn2a, Ki67 ) %>%
  summarise( tot = sum( Freq ) ) %>%
  left_join( outerCircleData1 %>% select(Epithtype, Cdkn2a, tot_cells) ) %>%
  ungroup() %>%
  mutate(
    ymax = cumsum(tot),
    ymin = lag( ymax, n = 1, default = 0 )
  ) 

outerCircleData3 <- dt %>%
  group_by(Epithtype, Cdkn2a,  Ki67, EMstate ) %>%
  summarise( tot = sum( Freq ) ) %>%
  left_join( outerCircleData2 %>% select(Epithtype,Cdkn2a, Ki67, tot_cells) ) %>%
  ungroup() %>%
  mutate(
    ymax = cumsum(tot),
    ymin = lag( ymax, n = 1, default = 0 )
  )

outerCircleData4 <- dt %>%
  group_by(Epithtype, Cdkn2a,  Ki67, EMstate, Vim ) %>%
  summarise( tot = sum( Freq ) ) %>%
  left_join( outerCircleData3 %>% select(Epithtype,Cdkn2a, Ki67,EMstate, tot_cells) ) %>%
  ungroup() %>%
  mutate(
    ymax = cumsum(tot),
    ymin = lag( ymax, n = 1, default = 0 )
  )

innerCircle <- ggplot( rectdata ) +
  geom_rect(
    aes( xmin = 2, xmax = 3, ymin = ymin, ymax = ymax, fill = Epithtype ),
    color = "black"
  ) 
         

outerCircle1 <-
  geom_rect(
    data = outerCircleData1,
    aes( xmin = 3, xmax = 4, ymin = ymin, ymax = ymax, fill = Cdkn2a ),
    color = "black" ) 

outerCircle2 <-
  geom_rect(
    data = outerCircleData2,
    aes( xmin = 4, xmax = 5, ymin = ymin, ymax = ymax, fill = Ki67 ),
    color = "black" ) 

outerCircle3 <-
  geom_rect(
    data = outerCircleData3,
    aes( xmin = 5, xmax = 6, ymin = ymin, ymax = ymax, fill = EMstate ),
    color = "black")    

outerCircle4 <-
  geom_rect(
    data = outerCircleData4,
    aes( xmin = 6, xmax = 7, ymin = ymin, ymax = ymax, fill = Vim ),
    color = "black") 



name=unique(dt$Parent)
p <- innerCircle  + outerCircle1 + outerCircle2 + outerCircle3 +outerCircle4+
 scale_fill_manual(values= c("Basal"= "#46D459","Krt5Krt8p"= "#D076FA","Luminal"= "#E67373","none"= "#E6DADA","p16neg"="#F7D2F7","p16pos" = "#F08DE0","Ki67neg"= "#839BDE","Ki67pos"= "#3C6CF0","Vimneg" ="#F5F3DC","Mesenchymal"="#C9F3F7","Epithelial"= "#DAE9F0","undetermined"= "#E6E6E6", "DP"="#76E0F0",   "Vimpos"="#FAFA7F" ))  + ggtitle(name) + themplot


g= p +  coord_polar( theta = "y" ) + themplot

pdf(paste0(PLOTdir, "/New_ggSunburst_ggBarplot_", name, ".pdf"))
print(p)
print(g)
dev.off()
}

for(i in (unique(dtest$Parent)[-1])){
  dt= dtest %>% filter(Parent== i)
  sunburst_ggplot_multiplelayers(dt)
}

dt= dtest %>% filter(Parent=="CreN")
cren=mat %>% dplyr::filter(Parent=="CreN")
dt2= dtest %>% dplyr::filter(Parent=="CreN")%>% dplyr::filter(Epithtype=="Luminal")

dt2= dtest %>% dplyr::filter(Parent=="Juxta")%>% dplyr::filter(Cdkn2a=="p16neg")
%>% dplyr::filter(EMstate=="DP")

dt3= dtest %>% dplyr::filter(Parent=="Juxta")%>% dplyr::filter(Cdkn2a=="p16neg")%>% dplyr::filter(EMstate=="Epithelial")%>% dplyr::filter(Vim=="Vimpos")


dt4=dtest %>% dplyr::filter(Parent=="Juxta")%>% dplyr::filter(Cdkn2a=="p16neg")
sum(dt4$Freq)

%>% dplyr::filter(Epithtype=="Luminal")



```

New sunburst:
```{r}
juxta= mat %>% dplyr::filter(Parent=="Juxta") %>% dplyr::filter(Cdkn2a==1 & Krt8==1 & Krt5==0)
epith=juxta[,-c(1,9:17)]

epith$Cdkn2a= ifelse(epith$Cdkn2a== 0, "p16neg", "p16pos")

epith$Krt8= ifelse(epith$Krt8== 0, "Krt8neg", "Krt8pos")

epith$Vim= ifelse(epith$Vim== 0, "Vimneg", "Vimpos")

epith$Ecad= ifelse(epith$Ecad== 0, "Ecadneg", "Ecadpos")


dt= as.data.frame(table(epith))

rectdata <- dt %>%
  group_by( Krt8 ) %>%
  summarise( tot_cells = sum( Freq ) ) %>%
  ungroup() %>%
  mutate(
    ymax = cumsum( tot_cells ),
    ymin = lag( ymax, n = 1, default = 0 )
  )


outerCircleData1 <- dt %>%
  group_by( Krt8, Cdkn2a ) %>%
  summarise( tot = sum( Freq ) ) %>%
  left_join( rectdata %>% select( Krt8,tot_cells) ) %>%
  ungroup() %>%
  mutate(
    ymax = cumsum(tot),
    ymin = lag( ymax, n = 1, default = 0 )
  ) 

outerCircleData2 <- dt %>%
  group_by( Krt8, Cdkn2a, Ecad ) %>%
  summarise( tot = sum( Freq ) ) %>%
  left_join( outerCircleData1 %>% select(Krt8, Cdkn2a, tot_cells) ) %>%
  ungroup() %>%
  mutate(
    ymax = cumsum(tot),
    ymin = lag( ymax, n = 1, default = 0 )
  ) 

outerCircleData3 <- dt %>%
  group_by(Krt8, Cdkn2a,  Ecad, Vim ) %>%
  summarise( tot = sum( Freq ) ) %>%
  left_join( outerCircleData2 %>% select(Krt8,Cdkn2a, Ecad, tot_cells) ) %>%
  ungroup() %>%
  mutate(
    ymax = cumsum(tot),
    ymin = lag( ymax, n = 1, default = 0 )
  )

outerCircleData4 <- dt %>%
  group_by(Krt8, Cdkn2a,  Ecad ) %>%
  summarise( tot = sum( Freq ) ) %>%
  left_join( outerCircleData3 %>% select(Krt8,Cdkn2a, Ecad, tot_cells) ) %>%
  ungroup() %>%
  mutate(
    ymax = cumsum(tot),
    ymin = lag( ymax, n = 1, default = 0 )
  )

innerCircle <- ggplot( rectdata ) +
  geom_rect(
    aes( xmin = 2, xmax = 3, ymin = ymin, ymax = ymax, fill = Krt8 ),
    color = "black"
  ) 
         

outerCircle1 <-
  geom_rect(
    data = outerCircleData1,
    aes( xmin = 3, xmax = 4, ymin = ymin, ymax = ymax, fill = Cdkn2a ),
    color = "black" ) 

outerCircle2 <-
  geom_rect(
    data = outerCircleData2,
    aes( xmin = 4, xmax = 5, ymin = ymin, ymax = ymax, fill = Ecad ),
    color = "black" ) 

outerCircle3 <-
  geom_rect(
    data = outerCircleData3,
    aes( xmin = 5, xmax = 6, ymin = ymin, ymax = ymax, fill = Vim ),
    color = "black")    

outerCircle4 <-
  geom_rect(
    data = outerCircleData4,
    aes( xmin = 6, xmax = 7, ymin = ymin, ymax = ymax, fill = Vim ),
    color = "black") 



name=unique(dt$Parent)
p <- innerCircle  + outerCircle1 + outerCircle2 + outerCircle3 +
 scale_fill_manual(values= c("Krt8pos"= "#E67373","Krt8neg"= "#E6DADA","p16neg"="#F7D2F7","p16pos" = "#F08DE0","Vimneg" ="#F5F3DC","Vimpos"="#FAFA7F","Ecadpos"="#C9F3F7","Ecadneg"= "#92999E" ))  + ggtitle(name) + themplot


g= p +  coord_polar( theta = "y" ) + themplot

pdf(file.path(figDir, "Sunburst_Final_Luminal_p16Neg_Juxta.pdf"))
print(g)
dev.off()
```







Control Cre3M:

```{r}
epith$Parent= ifelse(epith$Parent=="Duct_Juxta", "Duct", epith$Parent)

dtest= as.data.frame(table(epith))

dt= dtest %>% filter(Parent== "Duct" & Epithtype=="Luminal" & Cdkn2a=="p16neg" & name=="Cre_neg_5M_Duct")

rectdata <- dt %>%
  group_by( Epithtype ) %>%
  summarise( tot_cells = sum( Freq ) ) %>%
  ungroup() %>%
  mutate(
    ymax = cumsum( tot_cells ),
    ymin = lag( ymax, n = 1, default = 0 ) )

outerCircleData2 <- dt %>%
  group_by( Epithtype,  Ki67 ) %>%
  summarise( tot = sum( Freq ) ) %>%
  left_join( rectdata %>% select(Epithtype,  tot_cells) ) %>%
  ungroup() %>%
  mutate(
    ymax = cumsum(tot),
    ymin = lag( ymax, n = 1, default = 0 )
  ) 

outerCircleData3 <- dt %>%
  group_by(Epithtype,   Ki67, EMstate ) %>%
  summarise( tot = sum( Freq ) ) %>%
  left_join( outerCircleData2 %>% select(Epithtype, Ki67, tot_cells) ) %>%
  ungroup() %>%
  mutate(
    ymax = cumsum(tot),
    ymin = lag( ymax, n = 1, default = 0 )
  )

outerCircleData4 <- dt %>%
  group_by(Epithtype,   Ki67, EMstate, Vim ) %>%
  summarise( tot = sum( Freq ) ) %>%
  left_join( outerCircleData3 %>% select(Epithtype, Ki67,EMstate, tot_cells) ) %>%
  ungroup() %>%
  mutate(
    ymax = cumsum(tot),
    ymin = lag( ymax, n = 1, default = 0 )
  )

innerCircle <- ggplot( rectdata ) +
  geom_rect(
    aes( xmin = 2, xmax = 3, ymin = ymin, ymax = ymax, fill = Epithtype ),
    color = "black"
  ) 
        
outerCircle2 <-
  geom_rect(
    data = outerCircleData2,
    aes( xmin = 3, xmax = 4, ymin = ymin, ymax = ymax, fill = Ki67 ),
    color = "black" ) 

outerCircle3 <-
  geom_rect(
    data = outerCircleData3,
    aes( xmin = 4, xmax = 5, ymin = ymin, ymax = ymax, fill = EMstate ),
    color = "black")    

outerCircle4 <-
  geom_rect(
    data = outerCircleData4,
    aes( xmin = 5, xmax = 6, ymin = ymin, ymax = ymax, fill = Vim ),
    color = "black") 

name=unique(dt$name)

pdf(paste0(PLOTdir, "/Ggsunburst_P16Neg_Luminal_Duct_Creneg5M.pdf"))
p <- innerCircle  +  outerCircle2 + outerCircle3 +outerCircle4+
 scale_fill_manual(values= c("Basal"= "#46D459","Krt5Krt8p"= "#D076FA","Luminal"= "#E67373","none"= "#E6DADA","p16neg"="#F7D2F7","p16pos" = "#F08DE0","Ki67neg"= "#839BDE","Ki67pos"= "#3C6CF0","Vimneg" ="#F5F3DC","Mesenchymal"="#C9F3F7","Epithelial"= "#92999E","undetermined"= "#E6E6E6", "DP"="#76E0F0",   "Vimpos"="#FAFA7F" ))  + ggtitle("Duct_P16negative") + themplot
print(p)

g= p +  coord_polar( theta = "y" ) + themplot
print(g)
dev.off()
  

```


```{r}
sunburst_ggplot_multiplelayers_V2= function(dt){

#dt= dt %>% filter(Cdkn2a=="p16pos")
dt= dt %>% filter(Cdkn2a=="p16pos")

rectdata <- dt %>%
  group_by( Epithtype ) %>%
  summarise( tot_cells = sum( Freq ) ) %>%
  ungroup() %>%
  mutate(
    ymax = cumsum( tot_cells ),
    ymin = lag( ymax, n = 1, default = 0 ) )

outerCircleData2 <- dt %>%
  group_by( Epithtype,  Ki67 ) %>%
  summarise( tot = sum( Freq ) ) %>%
  left_join( rectdata %>% select(Epithtype,  tot_cells) ) %>%
  ungroup() %>%
  mutate(
    ymax = cumsum(tot),
    ymin = lag( ymax, n = 1, default = 0 )
  ) 

outerCircleData3 <- dt %>%
  group_by(Epithtype,   Ki67, EMstate ) %>%
  summarise( tot = sum( Freq ) ) %>%
  left_join( outerCircleData2 %>% select(Epithtype, Ki67, tot_cells) ) %>%
  ungroup() %>%
  mutate(
    ymax = cumsum(tot),
    ymin = lag( ymax, n = 1, default = 0 )
  )

outerCircleData4 <- dt %>%
  group_by(Epithtype,   Ki67, EMstate, Vim ) %>%
  summarise( tot = sum( Freq ) ) %>%
  left_join( outerCircleData3 %>% select(Epithtype, Ki67,EMstate, tot_cells) ) %>%
  ungroup() %>%
  mutate(
    ymax = cumsum(tot),
    ymin = lag( ymax, n = 1, default = 0 )
  )

innerCircle <- ggplot( rectdata ) +
  geom_rect(
    aes( xmin = 2, xmax = 3, ymin = ymin, ymax = ymax, fill = Epithtype ),
    color = "black"
  ) 
        
outerCircle2 <-
  geom_rect(
    data = outerCircleData2,
    aes( xmin = 3, xmax = 4, ymin = ymin, ymax = ymax, fill = Ki67 ),
    color = "black" ) 

outerCircle3 <-
  geom_rect(
    data = outerCircleData3,
    aes( xmin = 4, xmax = 5, ymin = ymin, ymax = ymax, fill = EMstate ),
    color = "black")    

outerCircle4 <-
  geom_rect(
    data = outerCircleData4,
    aes( xmin = 5, xmax = 6, ymin = ymin, ymax = ymax, fill = Vim ),
    color = "black") 

name=unique(dt$Parent)

pdf(paste0(figDir, "/New_Ggsunburst_P16PosNeg_NewEMstate_", name ,".pdf"))
p <- innerCircle  +  outerCircle2 + outerCircle3 +outerCircle4+
 scale_fill_manual(values= c("Basal"= "#46D459","Krt5Krt8p"= "#D076FA","Luminal"= "#E67373","none"= "#E6DADA","p16neg"="#F7D2F7","p16pos" = "#F08DE0","Ki67neg"= "#839BDE","Ki67pos"= "#3C6CF0","Vimneg" ="#F5F3DC","Mesenchymal"="#C9F3F7","Epithelial"= "#92999E","undetermined"= "#E6E6E6", "DP"="#76E0F0",   "Vimpos"="#FAFA7F" ))  + ggtitle(paste0(name,"_P16positive")) + themplot
print(p)

g= p +  coord_polar( theta = "y" ) + themplot
print(g)
dev.off()
  
}

epith$Parent= ifelse(epith$Parent=="Duct_Juxta", "Duct", epith$Parent)
dtest= as.data.frame(table(epith))


for(i in unique(dtest$Parent)[-1]){
  dt= dtest %>% filter(Parent== i)
  sunburst_ggplot_multiplelayers_V2(dt)
}

 dt= dtest %>% filter(name== "Cre_pos_3M_Duct")
  sunburst_ggplot_multiplelayers_V2(dt)
```



```{r}
sunburst_ggplot_multiplelayers= function(dt){
  dt= dt %>% filter(Cdkn2a=="p16neg")
  
rectdata <- dt %>%
  group_by( Epithtype ) %>%
  summarise( tot_cells = sum( Freq ) ) %>%
  ungroup() %>%
  mutate(
    ymax = cumsum( tot_cells ),
    ymin = lag( ymax, n = 1, default = 0 ) )

outerCircleData2 <- dt %>%
  group_by( Epithtype,  Ki67 ) %>%
  summarise( tot = sum( Freq ) ) %>%
  left_join( rectdata %>% select(Epithtype,  tot_cells) ) %>%
  ungroup() %>%
  mutate(
    ymax = cumsum(tot),
    ymin = lag( ymax, n = 1, default = 0 )
  ) 

outerCircleData3 <- dt %>%
  group_by(Epithtype,   Ki67, EMstate ) %>%
  summarise( tot = sum( Freq ) ) %>%
  left_join( outerCircleData2 %>% select(Epithtype, Ki67, tot_cells) ) %>%
  ungroup() %>%
  mutate(
    ymax = cumsum(tot),
    ymin = lag( ymax, n = 1, default = 0 )
  )



innerCircle <- ggplot( rectdata ) +
  geom_rect(
    aes( xmin = 2, xmax = 3, ymin = ymin, ymax = ymax, fill = Epithtype ),
    color = "black"
  ) 
        
outerCircle2 <-
  geom_rect(
    data = outerCircleData2,
    aes( xmin = 3, xmax = 4, ymin = ymin, ymax = ymax, fill = Ki67 ),
    color = "black" ) 

outerCircle3 <-
  geom_rect(
    data = outerCircleData3,
    aes( xmin = 4, xmax = 5, ymin = ymin, ymax = ymax, fill = EMstate ),
    color = "black")    



name=unique(dt$Parent)
p <- innerCircle  +  outerCircle2 + outerCircle3 +
 scale_fill_manual(values= c("Basal"= "#46D459","Krt5Krt8p"= "#D076FA","Luminal"= "#E67373","none"= "#E6DADA","p16neg"="#F7D2F7","p16pos" = "#F08DE0","Ki67neg"= "#839BDE","Ki67pos"= "#3C6CF0","Vimneg" ="#F5F3DC","Mesenchymal"="#C9F3F7","Epithelial"= "#DAE9F0","undetermined"= "#E6E6E6", "DP"="#76E0F0",   "Vimpos"="#FAFA7F" ))  + ggtitle(name) + themplot


g= p +  coord_polar( theta = "y" ) + themplot

pdf(paste0(figDir, "/Sunburst_NewEMstate_p16Neg_", name, ".pdf"))
print(p)
print(g)
dev.off()
}

for(i in (unique(dtest$Parent)[-1])){
  dt= dtest %>% filter(Parent== i)
  sunburst_ggplot_multiplelayers(dt)
}

qs::qsave(dtest, file.path(figDir, "Table_multistaining_States.qs"))

```
